5.0: Wang-Tsiatis Bounds

gsDesign offers the option of using Wang-Tsiatis bounds as an alternative to the spending function approach to group sequential design. Wang-Tsiatis bounds include both Pocock and O'Brien-Fleming designs. Wang-Tsiatis bounds are currently only available for 1-sided and symmetric 2-sided designs. Wang-Tsiatis bounds are typically used with equally spaced timing between analyses, but the option is available to use them with unequal spacing.

WT(d, alpha, a, timing, tol = 1e-06, r = 18)

Arguments

d
FUNCTION_PARAMETER

alpha
FUNCTION_PARAMETER

a
FUNCTION_PARAMETER

timing
FUNCTION_PARAMETER

tol
FUNCTION_PARAMETER

r
FUNCTION_PARAMETER

Details

Wang-Tsiatis bounds are defined as follows. Assume \(k\) analyses and let \(Z_i\) represent the upper bound and \(t_i\) the proportion of the total planned sample size for the \(i\)-th analysis, \(i=1,2,\ldots,k\). Let \(\Delta\) be a real-value. Typically \(\Delta\) will range from 0 (O'Brien-Fleming design) to 0.5 (Pocock design). The upper boundary is defined by $$ct_i^{\Delta-0.5}$$ for \(i= 1,2,\ldots,k\) where \(c\) depends on the other parameters. The parameter \(\Delta\) is supplied to gsDesign() in the parameter sfupar. For O'Brien-Fleming and Pocock designs there is also a calling sequence that does not require a parameter. See examples.

Note

The manual is not linked to this help file, but is available in library/gsdesign/doc/gsDesignManual.pdf in the directory where R is installed.

References

Jennison C and Turnbull BW (2000), Group Sequential Methods with Applications to Clinical Trials. Boca Raton: Chapman and Hall.

See also

Examples

# Pocock design
gsDesign(test.type = 2, sfu = "Pocock")

#> Symmetric two-sided group sequential design with
#> 90 % power and 2.5 % Type I Error.
#> Spending computations assume trial stops
#> if a bound is crossed.
#> 
#>            Sample
#>             Size 
#>   Analysis Ratio*  Z   Nominal p  Spend
#>          1  0.384 2.29     0.011 0.0110
#>          2  0.767 2.29     0.011 0.0079
#>          3  1.151 2.29     0.011 0.0060
#>      Total                       0.0250 
#> 
#> ++ alpha spending:
#>  Pocock boundary.
#> * Sample size ratio compared to fixed design with no interim
#> 
#> Boundary crossing probabilities and expected sample size
#> assume any cross stops the trial
#> 
#> Upper boundary (power or Type I Error)
#>           Analysis
#>    Theta     1      2      3 Total   E{N}
#>   0.0000 0.011 0.0079 0.0060 0.025 1.1276
#>   3.2415 0.389 0.3421 0.1689 0.900 0.7210
#> 
#> Lower boundary (futility or Type II Error)
#>           Analysis
#>    Theta     1      2     3 Total
#>   0.0000 0.011 0.0079 0.006 0.025
#>   3.2415 0.000 0.0000 0.000 0.000

# alternate call to get Pocock design specified using
# Wang-Tsiatis option and Delta=0.5
gsDesign(test.type = 2, sfu = "WT", sfupar = 0.5)

#> Symmetric two-sided group sequential design with
#> 90 % power and 2.5 % Type I Error.
#> Spending computations assume trial stops
#> if a bound is crossed.
#> 
#>            Sample
#>             Size 
#>   Analysis Ratio*  Z   Nominal p  Spend
#>          1  0.384 2.29     0.011 0.0110
#>          2  0.767 2.29     0.011 0.0079
#>          3  1.151 2.29     0.011 0.0060
#>      Total                       0.0250 
#> 
#> ++ alpha spending:
#>  Wang-Tsiatis boundary with Delta = 0.5.
#> * Sample size ratio compared to fixed design with no interim
#> 
#> Boundary crossing probabilities and expected sample size
#> assume any cross stops the trial
#> 
#> Upper boundary (power or Type I Error)
#>           Analysis
#>    Theta     1      2      3 Total   E{N}
#>   0.0000 0.011 0.0079 0.0060 0.025 1.1276
#>   3.2415 0.389 0.3421 0.1689 0.900 0.7210
#> 
#> Lower boundary (futility or Type II Error)
#>           Analysis
#>    Theta     1      2     3 Total
#>   0.0000 0.011 0.0079 0.006 0.025
#>   3.2415 0.000 0.0000 0.000 0.000

# this is how this might work with a spending function approach
# Hwang-Shih-DeCani spending function with gamma=1 is often used
# to approximate Pocock design
gsDesign(test.type = 2, sfu = sfHSD, sfupar = 1)

#> Symmetric two-sided group sequential design with
#> 90 % power and 2.5 % Type I Error.
#> Spending computations assume trial stops
#> if a bound is crossed.
#> 
#>            Sample
#>             Size 
#>   Analysis Ratio*  Z   Nominal p  Spend
#>          1  0.385 2.28    0.0112 0.0112
#>          2  0.771 2.28    0.0112 0.0080
#>          3  1.156 2.30    0.0107 0.0058
#>      Total                       0.0250 
#> 
#> ++ alpha spending:
#>  Hwang-Shih-DeCani spending function with gamma = 1.
#> * Sample size ratio compared to fixed design with no interim
#> 
#> Boundary crossing probabilities and expected sample size
#> assume any cross stops the trial
#> 
#> Upper boundary (power or Type I Error)
#>           Analysis
#>    Theta      1      2      3 Total   E{N}
#>   0.0000 0.0112 0.0080 0.0058 0.025 1.1327
#>   3.2415 0.3933 0.3418 0.1649 0.900 0.7213
#> 
#> Lower boundary (futility or Type II Error)
#>           Analysis
#>    Theta      1     2      3 Total
#>   0.0000 0.0112 0.008 0.0058 0.025
#>   3.2415 0.0000 0.000 0.0000 0.000

# unequal spacing works,  but may not be desirable
gsDesign(test.type = 2, sfu = "Pocock", timing = c(.1, .2))

#> Symmetric two-sided group sequential design with
#> 90 % power and 2.5 % Type I Error.
#> Spending computations assume trial stops
#> if a bound is crossed.
#> 
#>            Sample
#>             Size 
#>   Analysis Ratio*  Z   Nominal p  Spend
#>          1  0.122 2.34    0.0096 0.0096
#>          2  0.244 2.34    0.0096 0.0070
#>          3  1.221 2.34    0.0096 0.0084
#>      Total                       0.0250 
#> 
#> ++ alpha spending:
#>  Pocock boundary.
#> * Sample size ratio compared to fixed design with no interim
#> 
#> Boundary crossing probabilities and expected sample size
#> assume any cross stops the trial
#> 
#> Upper boundary (power or Type I Error)
#>           Analysis
#>    Theta      1      2      3 Total   E{N}
#>   0.0000 0.0096 0.0070 0.0084 0.025 1.1864
#>   3.2415 0.1135 0.1479 0.6386 0.900 0.9517
#> 
#> Lower boundary (futility or Type II Error)
#>           Analysis
#>    Theta      1     2      3  Total
#>   0.0000 0.0096 0.007 0.0084 0.0250
#>   3.2415 0.0003 0.000 0.0000 0.0003

# spending function approximation to Pocock with unequal spacing
# is quite different from this
gsDesign(test.type = 2, sfu = sfHSD, sfupar = 1, timing = c(.1, .2))

#> Symmetric two-sided group sequential design with
#> 90 % power and 2.5 % Type I Error.
#> Spending computations assume trial stops
#> if a bound is crossed.
#> 
#>            Sample
#>             Size 
#>   Analysis Ratio*  Z   Nominal p  Spend
#>          1  0.106 2.67    0.0038 0.0038
#>          2  0.213 2.63    0.0043 0.0034
#>          3  1.064 2.08    0.0189 0.0178
#>      Total                       0.0250 
#> 
#> ++ alpha spending:
#>  Hwang-Shih-DeCani spending function with gamma = 1.
#> * Sample size ratio compared to fixed design with no interim
#> 
#> Boundary crossing probabilities and expected sample size
#> assume any cross stops the trial
#> 
#> Upper boundary (power or Type I Error)
#>           Analysis
#>    Theta      1      2      3 Total   E{N}
#>   0.0000 0.0038 0.0034 0.0178 0.025 1.0507
#>   3.2415 0.0531 0.0939 0.7530 0.900 0.9328
#> 
#> Lower boundary (futility or Type II Error)
#>           Analysis
#>    Theta      1      2      3  Total
#>   0.0000 0.0038 0.0034 0.0178 0.0250
#>   3.2415 0.0001 0.0000 0.0000 0.0001

# One-sided O'Brien-Fleming design
gsDesign(test.type = 1, sfu = "OF")

#> One-sided group sequential design with
#> 90 % power and 2.5 % Type I Error.
#>            Sample
#>             Size 
#>   Analysis Ratio*  Z   Nominal p  Spend
#>          1  0.339 3.47    0.0003 0.0003
#>          2  0.677 2.45    0.0071 0.0069
#>          3  1.016 2.00    0.0225 0.0178
#>      Total                       0.0250 
#> 
#> ++ alpha spending:
#>  O'Brien-Fleming boundary.
#> * Sample size ratio compared to fixed design with no interim
#> 
#> Boundary crossing probabilities and expected sample size
#> assume any cross stops the trial
#> 
#> Upper boundary (power or Type I Error)
#>           Analysis
#>    Theta      1      2      3 Total   E{N}
#>   0.0000 0.0003 0.0069 0.0178 0.025 1.0136
#>   3.2415 0.0565 0.5288 0.3147 0.900 0.7987

# alternate call to get O'Brien-Fleming design specified using
# Wang-Tsiatis option and Delta=0
gsDesign(test.type = 1, sfu = "WT", sfupar = 0)

#> One-sided group sequential design with
#> 90 % power and 2.5 % Type I Error.
#>            Sample
#>             Size 
#>   Analysis Ratio*  Z   Nominal p  Spend
#>          1  0.339 3.47    0.0003 0.0003
#>          2  0.677 2.45    0.0071 0.0069
#>          3  1.016 2.00    0.0225 0.0178
#>      Total                       0.0250 
#> 
#> ++ alpha spending:
#>  Wang-Tsiatis boundary with Delta = 0.
#> * Sample size ratio compared to fixed design with no interim
#> 
#> Boundary crossing probabilities and expected sample size
#> assume any cross stops the trial
#> 
#> Upper boundary (power or Type I Error)
#>           Analysis
#>    Theta      1      2      3 Total   E{N}
#>   0.0000 0.0003 0.0069 0.0178 0.025 1.0136
#>   3.2415 0.0565 0.5288 0.3147 0.900 0.7987

# WT function [sinew] ----