Create a summary table

Create a tibble to summarize an object; currently only implemented for gsBinomialExact.

Usage

as_table(x, ...)

# S3 method for gsBinomialExact
as_table(x, ...)

Arguments

x

Object to be summarized.

...

Other parameters that may be specific to the object.

Value

A tibble which may have an extended class to enable further output processing.

Details

Currently only implemented for gsBinomialExact objects. Creates a table to summarize an object. For gsBinomialExact, this summarized operating characteristics across a range of effect sizes.

See also

Examples

b <- binomialSPRT(p0 = .1, p1 = .35, alpha = .08, beta = .2, minn = 10, maxn = 25)
b_power <- gsBinomialExact(
  k = length(b$n.I), theta = seq(.1, .45, .05), n.I = b$n.I,
  a = b$lower$bound, b = b$upper$bound
)
b_power %>%
  as_table() %>%
  as_gt()
#> <div id="ijvqnpahfe" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>#ijvqnpahfe table {
#>   font-family: system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
#>   -webkit-font-smoothing: antialiased;
#>   -moz-osx-font-smoothing: grayscale;
#> }
#> 
#> #ijvqnpahfe thead, #ijvqnpahfe tbody, #ijvqnpahfe tfoot, #ijvqnpahfe tr, #ijvqnpahfe td, #ijvqnpahfe th {
#>   border-style: none;
#> }
#> 
#> #ijvqnpahfe p {
#>   margin: 0;
#>   padding: 0;
#> }
#> 
#> #ijvqnpahfe .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   line-height: normal;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #ijvqnpahfe .gt_caption {
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #ijvqnpahfe .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #ijvqnpahfe .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 3px;
#>   padding-bottom: 5px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #ijvqnpahfe .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ijvqnpahfe .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #ijvqnpahfe .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ijvqnpahfe .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #ijvqnpahfe .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #ijvqnpahfe .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #ijvqnpahfe .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #ijvqnpahfe .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #ijvqnpahfe .gt_spanner_row {
#>   border-bottom-style: hidden;
#> }
#> 
#> #ijvqnpahfe .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   text-align: left;
#> }
#> 
#> #ijvqnpahfe .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #ijvqnpahfe .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #ijvqnpahfe .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #ijvqnpahfe .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #ijvqnpahfe .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ijvqnpahfe .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #ijvqnpahfe .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #ijvqnpahfe .gt_row_group_first th {
#>   border-top-width: 2px;
#> }
#> 
#> #ijvqnpahfe .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ijvqnpahfe .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #ijvqnpahfe .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #ijvqnpahfe .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #ijvqnpahfe .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ijvqnpahfe .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #ijvqnpahfe .gt_last_grand_summary_row_top {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: double;
#>   border-bottom-width: 6px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #ijvqnpahfe .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #ijvqnpahfe .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #ijvqnpahfe .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ijvqnpahfe .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ijvqnpahfe .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #ijvqnpahfe .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #ijvqnpahfe .gt_left {
#>   text-align: left;
#> }
#> 
#> #ijvqnpahfe .gt_center {
#>   text-align: center;
#> }
#> 
#> #ijvqnpahfe .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #ijvqnpahfe .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #ijvqnpahfe .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #ijvqnpahfe .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #ijvqnpahfe .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #ijvqnpahfe .gt_footnote_marks {
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#>   position: initial;
#> }
#> 
#> #ijvqnpahfe .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #ijvqnpahfe .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #ijvqnpahfe .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #ijvqnpahfe .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #ijvqnpahfe .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #ijvqnpahfe .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> </style>
#>   <table class="gt_table" data-quarto-disable-processing="false" data-quarto-bootstrap="false">
#>   <thead>
#>     <tr class="gt_heading">
#>       <td colspan="4" class="gt_heading gt_title gt_font_normal" style>Operating Characteristics for the Truncated SPRT Design</td>
#>     </tr>
#>     <tr class="gt_heading">
#>       <td colspan="4" class="gt_heading gt_subtitle gt_font_normal gt_bottom_border" style>Assumes trial evaluated sequentially after each response</td>
#>     </tr>
#>     <tr class="gt_col_headings gt_spanner_row">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="2" colspan="1" scope="col" id="Underlying&lt;br&gt;response rate">Underlying<br>response rate</th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="2" scope="colgroup" id="Probability of crossing">
#>         <span class="gt_column_spanner">Probability of crossing</span>
#>       </th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="2" colspan="1" scope="col" id="Average&lt;br&gt;sample size">Average<br>sample size</th>
#>     </tr>
#>     <tr class="gt_col_headings">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col" id="Futility bound">Futility bound</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1" scope="col" id="Efficacy bound">Efficacy bound</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td headers="theta" class="gt_row gt_right">10%</td>
#> <td headers="Lower" class="gt_row gt_right">0.94</td>
#> <td headers="Upper" class="gt_row gt_right">0.04</td>
#> <td headers="en" class="gt_row gt_right">12.1</td></tr>
#>     <tr><td headers="theta" class="gt_row gt_right">15%</td>
#> <td headers="Lower" class="gt_row gt_right">0.78</td>
#> <td headers="Upper" class="gt_row gt_right">0.15</td>
#> <td headers="en" class="gt_row gt_right">13.6</td></tr>
#>     <tr><td headers="theta" class="gt_row gt_right">20%</td>
#> <td headers="Lower" class="gt_row gt_right">0.57</td>
#> <td headers="Upper" class="gt_row gt_right">0.32</td>
#> <td headers="en" class="gt_row gt_right">14.3</td></tr>
#>     <tr><td headers="theta" class="gt_row gt_right">25%</td>
#> <td headers="Lower" class="gt_row gt_right">0.37</td>
#> <td headers="Upper" class="gt_row gt_right">0.53</td>
#> <td headers="en" class="gt_row gt_right">14.2</td></tr>
#>     <tr><td headers="theta" class="gt_row gt_right">30%</td>
#> <td headers="Lower" class="gt_row gt_right">0.22</td>
#> <td headers="Upper" class="gt_row gt_right">0.71</td>
#> <td headers="en" class="gt_row gt_right">13.4</td></tr>
#>     <tr><td headers="theta" class="gt_row gt_right">35%</td>
#> <td headers="Lower" class="gt_row gt_right">0.12</td>
#> <td headers="Upper" class="gt_row gt_right">0.84</td>
#> <td headers="en" class="gt_row gt_right">12.5</td></tr>
#>     <tr><td headers="theta" class="gt_row gt_right">40%</td>
#> <td headers="Lower" class="gt_row gt_right">0.06</td>
#> <td headers="Upper" class="gt_row gt_right">0.92</td>
#> <td headers="en" class="gt_row gt_right">11.6</td></tr>
#>     <tr><td headers="theta" class="gt_row gt_right">45%</td>
#> <td headers="Lower" class="gt_row gt_right">0.03</td>
#> <td headers="Upper" class="gt_row gt_right">0.97</td>
#> <td headers="en" class="gt_row gt_right">11.0</td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>