gsDesign: Group Sequential Design

Derives group sequential clinical trial designs and describes their properties. Particular focus on time-to-event, binary, and continuous outcomes. Largely based on methods described in Jennison, Christopher and Turnbull, Bruce W., 2000, "Group Sequential Methods with Applications to Clinical Trials" ISBN: 0-8493-0316-8.

See also

Author

Maintainer: Keaven Anderson keaven_anderson@merck.com

Other contributors:

Merck & Co., Inc., Rahway, NJ, USA and its affiliates (ROR) [copyright holder]