Create a summary table

Create a tibble to summarize an object; currently only implemented for gsBinomialExact.

Usage

as_table(x, ...)

# S3 method for class 'gsBinomialExact'
as_table(x, ...)

Arguments

x

Object to be summarized.

...

Other parameters that may be specific to the object.

Value

A tibble which may have an extended class to enable further output processing.

Details

Currently only implemented for gsBinomialExact objects. Creates a table to summarize an object. For gsBinomialExact, this summarized operating characteristics across a range of effect sizes.

See also

Examples

b <- binomialSPRT(p0 = .1, p1 = .35, alpha = .08, beta = .2, minn = 10, maxn = 25)
b_power <- gsBinomialExact(
  k = length(b$n.I), theta = seq(.1, .45, .05), n.I = b$n.I,
  a = b$lower$bound, b = b$upper$bound
)
if (interactive()) {
  b_power |>
    as_table() |>
    lt()
}